player spawning

+++ src/space-traders/Assets/Code/Infrastructure/Identifiers/IIdentifierService.cs
@@ -2,6 +2,6 @@
 {
     public interface IIdentifierService
     {
-        ulong Next();
+        uint Next();
     }
 }
